tests: event: disable an offending assertion in active-low event test
Kernel commit 223ecaf140b1 ("gpiolib: fix incorrect IRQ requesting of
an active-low lineevent") fixed an issue with line events not being
inverted with the GPIOHANDLE_REQUEST_ACTIVE_LOW flag. This change has
now been released in linux v5.2.7 and since we don't want to change
the minimum kernel version for a stable branch, the offending assertion
needs to be removed.

Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>